-merging OpenSER 0.9.4 with SER unstable
git-svn-id: https://openser.svn.sourceforge.net/svnroot/openser/trunk@30 689a6050-402a-0410-94f2-e92a70836424

+++ modules_k/jabber/xjab_worker.c
@@ -342,7 +342,7 @@ int xj_worker_process(xj_wlist jwl, char* jaddress, int jport, int rank,
 		if(!FD_ISSET(pipe, &mset))
 			goto step_y;
 		
-		if(read(pipe, &jsmsg, sizeof(jsmsg)) < sizeof(jsmsg))
+		if(read(pipe, &jsmsg, sizeof(jsmsg)) < (int)sizeof(jsmsg))
 		{
 			DBG("XJAB:xj_worker:%d: BROKEN PIPE - exiting\n", _xj_pid);
 			break;
